perf: Changed every variable to short and bytes (as they cannot exceed them), changed string to char[]
Changed variables in Time.cs to byte, changed AddTime() in Score.cs to accept byte variable, changed GenerateNumbers() in Program.cs to short with a and b being bytes. Changed string Name in Player.cs to char[]. All of these are for performance improvements in terms of memory usage.

+++ HumanCalculator/Score.cs
@@ -16,7 +16,7 @@ namespace HumanCalculator
         {
             this.Time = new Time();
         }
-        public void AddTime(double seconds)
+        public void AddTime(byte seconds)
         {
             this.Time.Times[currentLoop]=seconds;
             currentLoop++;
